requirements.txt
umap-learn==0.5.5
numpy==1.23.5
pandas==1.5.3
scikit-learn==1.2.2
prince==0.13.0
